refactor: 패키지 분리
chess > pieces
