+++ src/test/java/com/woopaca/javachess/chess/pieces/PawnTest.java
@@ -1,4 +1,4 @@
-package com.woopaca.javachess;
+package com.woopaca.javachess.chess.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
